# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for HRD Attendance System",
        "title": "HRD Attendance API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/attendance": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get attendance list with filter \u0026 pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendance"
                ],
                "summary": "Get All Attendance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Relations: user,tenant,setting",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/modelDto.AttendanceListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Record clock-in / clock-out with face \u0026 location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendance"
                ],
                "summary": "Record Attendance",
                "parameters": [
                    {
                        "description": "Attendance Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AttendanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AttendanceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/attendance/summary": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get summary of today's attendance with comparison",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendance"
                ],
                "summary": "Get Attendance Summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AttendanceSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login employee",
                "parameters": [
                    {
                        "description": "Login Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register new employee",
                "parameters": [
                    {
                        "description": "Register Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/email/test": {
            "post": {
                "description": "Send a test email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email"
                ],
                "summary": "Send Test Email",
                "parameters": [
                    {
                        "description": "Email Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.EmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.EmailRequest": {
            "type": "object",
            "required": [
                "html",
                "subject",
                "to"
            ],
            "properties": {
                "html": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "to": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.AttendanceAction": {
            "type": "string",
            "enum": [
                "clock_in",
                "clock_out"
            ],
            "x-enum-varnames": [
                "ClockIn",
                "ClockOut"
            ]
        },
        "model.AttendanceRequest": {
            "type": "object",
            "required": [
                "action",
                "latitude",
                "longitude"
            ],
            "properties": {
                "action": {
                    "enum": [
                        "clock_in",
                        "clock_out"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AttendanceAction"
                        }
                    ],
                    "example": "clock_in"
                },
                "latitude": {
                    "type": "number",
                    "example": -6.1339179
                },
                "longitude": {
                    "type": "number",
                    "example": 106.8329504
                },
                "media_url": {
                    "type": "string",
                    "example": "https://i.pinimg.com/control1/736x/41/e7/99/41e799436291fdfb4c969b80913a19fe.jpg"
                }
            }
        },
        "model.AttendanceResponse": {
            "type": "object",
            "properties": {
                "clock_in_latitude": {
                    "type": "number"
                },
                "clock_in_longitude": {
                    "type": "number"
                },
                "clock_in_media_url": {
                    "type": "string"
                },
                "clock_in_time": {
                    "type": "string"
                },
                "clock_out_latitude": {
                    "type": "number"
                },
                "clock_out_longitude": {
                    "type": "number"
                },
                "clock_out_media_url": {
                    "type": "string"
                },
                "clock_out_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.AttendanceStatus"
                },
                "user": {
                    "$ref": "#/definitions/model.UserResponse"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.AttendanceStatus": {
            "type": "string",
            "enum": [
                "working",
                "done",
                "late"
            ],
            "x-enum-varnames": [
                "StatusWorking",
                "StatusDone",
                "StatusLate"
            ]
        },
        "model.AttendanceSummaryResponse": {
            "type": "object",
            "properties": {
                "late_summary": {
                    "type": "integer"
                },
                "late_summary_diff": {
                    "type": "number"
                },
                "ontime_summary": {
                    "type": "integer"
                },
                "ontime_summary_diff": {
                    "type": "number"
                },
                "total_record": {
                    "type": "integer"
                },
                "total_record_diff": {
                    "type": "number"
                }
            }
        },
...
        "model.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@yopmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
...
        "model.RecentActivityResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "tenant_id"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Jl. Sudirman No. 1"
                },
                "department": {
                    "type": "string",
                    "example": "IT"
                },
                "email": {
                    "type": "string",
                    "example": "budi@company.com"
                },
                "name": {
                    "type": "string",
                    "example": "Budi Santoso"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "123456"
                },
                "phone_number": {
                    "type": "string",
                    "example": "08123456789"
                },
                "role_id": {
                    "type": "integer",
                    "example": 1
                },
                "tenant_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "model.RoleResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.TenantResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tenant_settings": {
                    "$ref": "#/definitions/model.TenantSetting"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Jl. Sudirman No. 1"
                },
                "attendances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Attendance"
                    }
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-05T12:00:00Z"
                },
                "department": {
                    "type": "string",
                    "example": "IT"
                },
                "email": {
                    "type": "string",
                    "example": "budi@company.com"
                },
                "employee_id": {
                    "type": "string",
                    "example": "FT-001"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "media_url": {
                    "type": "string",
                    "example": "https://cdn.example.com/profile/budi.jpg"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Budi Santoso"
                },
                "phone_number": {
                    "type": "string",
                    "example": "08123456789"
                },
                "recent_activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RecentActivity"
                    }
                },
                "role": {
                    "$ref": "#/definitions/model.Role"
                },
                "role_id": {
                    "type": "integer",
                    "example": 1
                },
                "tenant": {
                    "$ref": "#/definitions/model.Tenant"
                },
                "tenant_id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-04-05T12:30:00Z"
                }
            }
        },
...
        "model.UserResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Jl. Sudirman No. 1"
                },
                "attendances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AttendanceResponse"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "department": {
                    "type": "string",
                    "example": "IT"
                },
                "email": {
                    "type": "string",
                    "example": "budi@company.com"
                },
                "employee_id": {
                    "type": "string",
                    "example": "FT-001"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "media_url": {
                    "type": "string",
                    "example": "https://cdn.example.com/profile/budi.jpg"
                },
                "name": {
                    "type": "string",
                    "example": "Budi Santoso"
                },
                "phone_number": {
                    "type": "string",
                    "example": "08123456789"
                },
                "recent_activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RecentActivityResponse"
                    }
                },
                "role": {
                    "$ref": "#/definitions/model.RoleResponse"
                },
                "tenant": {
                    "$ref": "#/definitions/model.TenantResponse"
                },
                "tenant_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "modelDto.AttendanceListResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/modelDto.Meta"
                }
            }
        },
        "modelDto.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "modelDto.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}